Refactor ImageObject class to extend MediaObject and update related namespaces
