+++ src/Schema/Schema.php
@@ -2,7 +2,11 @@
 
 namespace Rami\SeoBundle\Schema;
 
+use Rami\SeoBundle\Schema\Intangible\Audience;
+use Rami\SeoBundle\Schema\Intangible\Service;
 use Rami\SeoBundle\Schema\Intangible\StructuredValue\ContactPoint\PostalAddress;
+use Rami\SeoBundle\Schema\Place\AdministrativeArea;
+use Rami\SeoBundle\Schema\Thing\Action;
 use Rami\SeoBundle\Schema\Thing\CreativeWork;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\Article;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\Article\SocialMediaPosting;
@@ -10,7 +14,8 @@ use Rami\SeoBundle\Schema\Thing\CreativeWork\Article\SocialMediaPosting\BlogPost
 use Rami\SeoBundle\Schema\Thing\CreativeWork\Blog;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\Course;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\EducationalOccupationalCredential;
-use Rami\SeoBundle\Schema\Thing\CreativeWork\ImageObject;
+use Rami\SeoBundle\Schema\Thing\CreativeWork\MediaObject\ImageObject;
+use Rami\SeoBundle\Schema\Thing\CreativeWork\TextObject;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\Thesis;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\WebPage;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\WebPage\AboutPage;
@@ -18,8 +23,7 @@ use Rami\SeoBundle\Schema\Thing\CreativeWork\WebPage\CollectionPage;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\WebPage\ContactPage;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\WebPageElement;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\WebSite;
-use Rami\SeoBundle\Schema\Thing\CreativeWork\TextObject;
-use Rami\SeoBundle\Schema\Thing\Action;
+use Rami\SeoBundle\Schema\Thing\Event;
 use Rami\SeoBundle\Schema\Thing\Intangible\BreadcrumbList;
 use Rami\SeoBundle\Schema\Thing\Intangible\DefinedTerm;
 use Rami\SeoBundle\Schema\Thing\Intangible\JobPosting;
@@ -29,18 +33,14 @@ use Rami\SeoBundle\Schema\Thing\Intangible\Specialty;
 use Rami\SeoBundle\Schema\Thing\Intangible\StructuredValue\MonetaryAmount;
 use Rami\SeoBundle\Schema\Thing\Intangible\StructuredValue\PropertyValue;
 use Rami\SeoBundle\Schema\Thing\Intangible\StructuredValue\SpeakableSpecification;
+use Rami\SeoBundle\Schema\Thing\Organization;
 use Rami\SeoBundle\Schema\Thing\Organization\LocalBusiness;
 use Rami\SeoBundle\Schema\Thing\Organization\LocalBusiness\Library;
 use Rami\SeoBundle\Schema\Thing\Organization\LocalBusiness\RadioStation;
 use Rami\SeoBundle\Schema\Thing\Organization\LocalBusiness\TravelAgency;
-use Rami\SeoBundle\Schema\Thing\Place\AdministrativeArea\Country;
-use Rami\SeoBundle\Schema\Intangible\Audience;
-use Rami\SeoBundle\Schema\Intangible\Service;
-use Rami\SeoBundle\Schema\Place\AdministrativeArea;
-use Rami\SeoBundle\Schema\Thing\Event;
-use Rami\SeoBundle\Schema\Thing\Organization;
 use Rami\SeoBundle\Schema\Thing\Person;
 use Rami\SeoBundle\Schema\Thing\Place;
+use Rami\SeoBundle\Schema\Thing\Place\AdministrativeArea\Country;
 
 class Schema implements SchemaInterface
 {
