+++ src/Schema/Traits/ThingTrait.php
@@ -5,7 +5,7 @@ namespace Rami\SeoBundle\Schema\Traits;
 use Rami\SeoBundle\Schema\Intangible\Service;
 use Rami\SeoBundle\Schema\Thing\Action;
 use Rami\SeoBundle\Schema\Thing\CreativeWork;
-use Rami\SeoBundle\Schema\Thing\CreativeWork\ImageObject;
+use Rami\SeoBundle\Schema\Thing\CreativeWork\MediaObject\ImageObject;
 use Rami\SeoBundle\Schema\Thing\CreativeWork\TextObject;
 use Rami\SeoBundle\Schema\Thing\Intangible\StructuredValue\PropertyValue;
 use Rami\SeoBundle\Schema\Thing\Organization;
